cef.patch from http://github.com/lindenlab/desura-app

+++ include/cef_capi.h
@@ -41,6 +41,8 @@
 extern "C" {
 #endif
 
+#include <inttypes.h>
+
 #include "internal/cef_export.h"
 #include "internal/cef_string.h"
 #include "internal/cef_string_list.h"
@@ -344,7 +346,7 @@ typedef struct _cef_base_t
 // Check that the structure |s|, which is defined with a cef_base_t member named
 // |base|, is large enough to contain the specified member |f|.
 #define CEF_MEMBER_EXISTS(s, f)   \
-  ((int)&((s)->f) - (int)(s) + sizeof((s)->f) <= (s)->base.size)
+  ((uintptr_t)&((s)->f) - (uintptr_t)(s) + sizeof((s)->f) <= (s)->base.size)
 
 #define CEF_MEMBER_MISSING(s, f)  (!CEF_MEMBER_EXISTS(s, f) || !((s)->f))
